pyproject.toml
[project]
name = "santa-elena-envio-full-set-a-recibido"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client",
    "google-auth",
    "google-auth-oauthlib",
    "openpyxl",
    "pandas",
    "py7zr",
    "pydantic",
    "pyyaml",
]

[project.scripts]
envio_full_set_a_recibido = "src.main:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.13",
]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
